perf(cli): detach CLI from desktop runtime
