+++ crates/vmux_cli/Cargo.toml
@@ -14,7 +14,7 @@ clap = { version = "4", features = ["derive"] }
 serde_json = { workspace = true }
 tokio = { workspace = true }
 vmux_mcp = { path = "../vmux_mcp" }
-vmux_service = { path = "../vmux_service" }
+vmux_service = { package = "vmux_service_client", path = "../vmux_service_client" }
 
 [dev-dependencies]
 assert_cmd = "2"
