+++ crates/vmux_core/Cargo.toml
@@ -9,6 +9,8 @@ publish = false
 rkyv = { workspace = true }
 serde = { workspace = true }
 uuid = { workspace = true }
+vmux_profile = { path = "../vmux_profile" }
+vmux_wire = { path = "../vmux_wire", features = ["bevy"] }
 
 [dev-dependencies]
 serde_json = { workspace = true }
