+++ crates/vmux_profile/build.rs
@@ -1,0 +1,6 @@
+#[path = "../build_git_env.rs"]
+mod build_git_env;
+
+fn main() {
+    build_git_env::emit();
+}
