+++ crates/vmux_service_client/src/lib.rs
@@ -1,0 +1,15 @@
+//! Lightweight access to the vmux background service.
+
+pub mod bundle;
+pub mod cli;
+pub mod client;
+pub mod framing;
+#[cfg(target_os = "macos")]
+pub mod launchd;
+pub mod paths;
+
+pub mod protocol {
+    pub use vmux_wire::protocol::*;
+}
+
+pub use paths::*;
