+++ crates/vmux_wire/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "vmux_wire"
+description = "Lightweight wire types shared by vmux processes"
+version.workspace = true
+edition.workspace = true
+publish = false
+
+[features]
+default = []
+bevy = ["dep:bevy_ecs", "dep:bevy_reflect"]
+
+[dependencies]
+rkyv = { workspace = true }
+serde = { workspace = true }
+uuid = { workspace = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+bevy_ecs = { workspace = true, optional = true }
+bevy_reflect = { workspace = true, optional = true }
+
+[dev-dependencies]
+serde_json = { workspace = true }
