+++ crates/vmux_wire/src/protocol/layout.rs
@@ -1,3 +1,3 @@
-pub use vmux_layout::protocol::{
+pub use crate::layout::{
     Focus, LayoutNode, LayoutSnapshot, NodeKind, SplitDirection, Stack, Tab, format_id, parse_id,
 };
